+++ tests/all.test.ts
@@ -1,0 +1,5 @@
+import "./api.test";
+import "./generation.test";
+import "./stats.test";
+import "./pages.test";
+import "./draw-sync.test";